ミドルウェアに移行
